CORE-7769 Move transaction verification classes to flow
