+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/transaction/UtxoTransactionVerification.kt
@@ -1,4 +1,4 @@
-package net.corda.ledger.utxo.data.transaction
+package net.corda.ledger.utxo.flow.impl.transaction
 
 import net.corda.v5.ledger.common.Party
 import net.corda.v5.ledger.common.transaction.TransactionMetadata
